+++ src/gekkouga/acmd/ground.rs
@@ -13,3 +13,5 @@ use smash::app::*;
 use smash::phx::Vector3f;
 use crate::util::*;
 use super::*;
+pub fn install() {
+}
