+++ src/gekkouga/frame.rs
@@ -13,3 +13,24 @@ use smash::app::*;
 use smash::phx::Vector3f;
 use crate::util::*;
 use super::*;
+pub fn install() {
+	smashline::install_agent_frames!(gren_frame);
+}
+#[fighter_frame( agent = FIGHTER_KIND_GEKKOUGA )]
+fn gren_frame(fighter: &mut L2CFighterCommon) {
+    unsafe {
+		let boma = smash::app::sv_system::battle_object_module_accessor(fighter.lua_state_agent);   
+		let status_kind = smash::app::lua_bind::StatusModule::status_kind(boma);
+		let ENTRY_ID = WorkModule::get_int(boma, *FIGHTER_INSTANCE_WORK_ID_INT_ENTRY_ID) as usize; 
+		let frame = MotionModule::frame(boma);
+		let end_frame = MotionModule::end_frame(boma);
+		if StatusModule::is_situation_changed(boma) && status_kind == *FIGHTER_STATUS_KIND_SPECIAL_LW {
+			if AttackModule::is_infliction_status(boma, *COLLISION_KIND_MASK_HIT) {
+				StatusModule::change_status_request_from_script(boma, *FIGHTER_STATUS_KIND_LANDING, true);
+			} else {
+				WorkModule::set_float(boma, 22.0, *FIGHTER_INSTANCE_WORK_ID_FLOAT_LANDING_FRAME);
+				StatusModule::change_status_request_from_script(boma, *FIGHTER_STATUS_KIND_LANDING_FALL_SPECIAL, true);
+			};
+		};
+	}
+}
